cleanup simple fixture
